Refactor TOC handling and improve serialization logic for enhanced performance and clarity
